MB-23664: Use max possible limit for Ephemeral backfill memory mgmt
In-memory backfilling is currently not memory managed. Therefore set the
scan buffer (per backfill buffer for managing backfill memory usage)
byte limit, item limit to max. Also, for the same reason as above set
the overall DCP backfill limit to max.

Unlike disk backfills, where there is luxury of surplus disk memory,
we cannot retain in-memory snapshots for long in case of a memory
pressure during backfill. This is because the snapshots themselves
use memory.

Memory management can be done, but will be considered in a future
task (MB-23734).

Reviewed-on: http://review.couchbase.org/76294
Tested-by: Build Bot <build@couchbase.com>

+++ src/dcp/backfill-manager.h
@@ -93,17 +93,17 @@ private:
 
     //! The scan buffer is for the current stream being backfilled
     struct {
-        uint32_t bytesRead;
-        uint32_t itemsRead;
-        uint32_t maxBytes;
-        uint32_t maxItems;
+        size_t bytesRead;
+        size_t itemsRead;
+        size_t maxBytes;
+        size_t maxItems;
     } scanBuffer;
 
     //! The buffer is the total bytes used by all backfills for this connection
     struct {
-        uint32_t bytesRead;
-        uint32_t maxBytes;
-        uint32_t nextReadSize;
+        size_t bytesRead;
+        size_t maxBytes;
+        size_t nextReadSize;
         bool full;
     } buffer;
 };
